+++ mast_aladin_lite/table/__init__.py
@@ -1,1 +1,1 @@
-from .table_widget import TableWidget
+from .mast_table import MastTable
